renamed package and unit test
